Add openSuse dependency for wget

+++ BUILDS/example6/build.sh
@@ -45,7 +45,14 @@ fi
 popd
 
 # cheating and adding a dependency for the build to work
-yum -q -y install gnutls-devel 
+if [ -f /usr/bin/yum ]; then
+  yum -q -y install gnutls-devel
+elif [ -f /usr/bin/zypper ]; then
+  zypper -q --non-interactive install libgnutls-devel
+else
+  echo "Unable to determine package manager"
+  exit 2
+fi
 
 echo ----------------------------------------------------
 echo Build the RPM
